+++ allauth/account/signals.py
@@ -9,8 +9,9 @@ password_set = Signal(providing_args=["request", "user"])
 password_changed = Signal(providing_args=["request", "user"])
 password_reset = Signal(providing_args=["request", "user"])
 
-email_confirmed = Signal(providing_args=["email_address"])
-email_confirmation_sent = Signal(providing_args=["confirmation"])
+email_confirmed = Signal(providing_args=["request", "email_address"])
+email_confirmation_sent = Signal(
+    providing_args=["request", "confirmation", "signup"])
 
 email_changed = Signal(
     providing_args=[